add: logout

+++ tunet_login.py
@@ -2,6 +2,7 @@ from hashlib import md5
 from binascii import hexlify
 import json
 import requests
+import sys
 from os import path
 
 
@@ -27,24 +28,32 @@ def login(username, password):
 
     print(r)
 
+def logout():
+    data = { 'action': 'logout' }
+    r = requests.post('http://net.tsinghua.edu.cn/do_login.php', data=data)
+    print(r)
+
 
 def main():
-    dir = path.dirname(__file__)
-    filename = path.join(dir, 'configure.json')
-    if (path.isfile(filename)):
-        with open(filename, 'r') as f:
-            data = json.load(f)
-            username = data['username']
-            password = data['password']
-    else:
-        username = input('username: ')
-        password = input('password: ')
-        autosave = input('save?(y/n): ')
-        if (autosave == 'y'):
-            with open(filename, 'w+') as f:
-                json.dump({ 'username': username, 'password': password }, f)
+    if len(sys.argv) == 1:
+        dir = path.dirname(__file__)
+        filename = path.join(dir, 'configure.json')
+        if (path.isfile(filename)):
+            with open(filename, 'r') as f:
+                data = json.load(f)
+                username = data['username']
+                password = data['password']
+        else:
+            username = input('username: ')
+            password = input('password: ')
+            autosave = input('save?(y/n): ')
+            if (autosave == 'y'):
+                with open(filename, 'w+') as f:
+                    json.dump({ 'username': username, 'password': password }, f)
 
-    login(username, password)
+        login(username, password)
+    elif sys.argv[1] == 'logout':
+        logout()
 
 if __name__ == '__main__':
     main()
